Moved authGW token verification from Productizer handlers to centralized service. Issuer is parsed from token for authGW verification.
