+++ VirtualFinland.UserAPI/src/VirtualFinland.UsersAPI/appsettings.Development.json
@@ -18,7 +18,7 @@
   "AuthGW":
   {
     "JwksJsonURL": "https://q88uo5prmh.execute-api.eu-north-1.amazonaws.com/auth/saml2/suomifi/.well-known/jwks.json",
-    "Issuer": "https://q88uo5prmh.execute-api.eu-north-1.amazonaws.com",
+    "Issuer": "virtual-finland/authentication-gw/suomifi",
     "AuthorizeURL": "https://q88uo5prmh.execute-api.eu-north-1.amazonaws.com/authorize"
   },
   "ConnectionStrings": {
