+++ VirtualFinland.UserAPI/src/VirtualFinland.UsersAPI/appsettings.Staging.json
@@ -18,7 +18,7 @@
   "AuthGW":
   {
     "JwksJsonURL": "https://a4j7hhtwlb.execute-api.eu-north-1.amazonaws.com/auth/saml2/suomifi/.well-known/jwks.json",
-    "Issuer": "https://a4j7hhtwlb.execute-api.eu-north-1.amazonaws.com",
+    "Issuer": "virtual-finland/authentication-gw/suomifi",
     "AuthorizeURL": "https://a4j7hhtwlb.execute-api.eu-north-1.amazonaws.com/authorize"
   },
   "ConnectionStrings": {
